fix linting error
